requirements.txt
Flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.1.1
Pillow==11.3.0
qrcode==8.2
Werkzeug==3.1.3
supabase==2.14.0
requests